refactor: Take spdx-validation package into use
Also remove spdxHelper from common-helpers package and move some unit
tests from clearance_ui app to spdx-validation package.
