+++ apps/clearance_ui/package.json
@@ -46,7 +46,6 @@
         "class-variance-authority": "0.7.0",
         "clsx": "2.0.0",
         "cmdk": "0.2.0",
-        "common-helpers": "*",
         "generate-password": "1.7.1",
         "js-yaml": "4.1.0",
         "lucide-react": "0.294.0",
@@ -58,6 +57,7 @@
         "react-hook-form": "7.48.2",
         "react-icons": "4.12.0",
         "react-resizable-panels": "^1.0.7",
+        "spdx-validation": "*",
         "tailwind-merge": "2.1.0",
         "tailwindcss-animate": "1.0.7",
         "validation-helpers": "*",
