+++ apps/clearance_ui/src/components/license_conclusions/ConclusionForm.tsx
@@ -7,9 +7,9 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useQueryClient } from "@tanstack/react-query";
 import { ZodiosResponseByPath } from "@zodios/core";
 import axios from "axios";
-import { parseSPDX } from "common-helpers";
 import { Info } from "lucide-react";
 import { useForm, useFormState } from "react-hook-form";
+import { parseSPDX } from "spdx-validation";
 import { userAPI } from "validation-helpers";
 import { z } from "zod";
 import { userHooks } from "@/hooks/zodiosHooks";
