+++ packages/common-helpers/index.d.ts
@@ -2,10 +2,6 @@
 //
 // SPDX-License-Identifier: MIT
 
-import parse from "spdx-expression-parse";
-
 declare const getCurrentDateTime: () => string;
 
-declare function parseSPDX(spdxString: string): parse.Info;
-
-export { getCurrentDateTime, parseSPDX };
+export { getCurrentDateTime };
