+++ packages/common-helpers/package.json
@@ -10,14 +10,12 @@
     },
     "dependencies": {
         "dotenv": "16.3.1",
-        "pako": "2.1.0",
-        "spdx-expression-parse": "github:doubleopen-project/spdx-expression-parse.js"
+        "pako": "2.1.0"
     },
     "devDependencies": {
         "@types/chai": "4.3.11",
         "@types/mocha": "10.0.6",
         "@types/pako": "2.0.3",
-        "@types/spdx-expression-parse": "^3.0.5",
         "chai": "4.3.10",
         "mocha": "10.2.0"
     }
