+++ packages/spdx-validation/package.json
@@ -1,7 +1,7 @@
 {
     "name": "spdx-validation",
-    "main": "./dist/index.js",
-    "module": "./dist/index.mjs",
+    "main": "./dist/src/index.js",
+    "module": "./dist/src/index.mjs",
     "types": "./index.d.ts",
     "scripts": {
         "build": "tsup && prettier --write dist/src/index.d.ts && (head -n 12 src/index.ts && cat dist/src/index.d.ts) > temp_file && mv temp_file index.d.ts",
