+++ packages/validation-helpers/package.json
@@ -10,7 +10,7 @@
     "dependencies": {
         "@zodios/express": "10.6.0",
         "check-password-strength": "2.0.7",
-        "common-helpers": "*",
+        "spdx-validation": "*",
         "is-glob": "4.0.3",
         "packageurl-js": "1.2.1",
         "zod": "3.22.4"
